requirements.txt
numpy==1.24.1
opencv-python==4.7.0.68
pyzmq==24.0.1
